requirements.txt
python-binance
streamlit
pandas
pandas-ta
plotly
streamlit-autorefresh